requirements.txt
python-dotenv
google-generativeai
mcp
pywinauto
Pillow
